Only allow group owner to delete invites

+++ app/Http/Controllers/GroupController.php
@@ -119,6 +119,11 @@ class GroupController extends Controller
 
     public function deleteInvite(Course $course, Group $group, GroupInvitation $groupInvitation): string
     {
+        $actor = $group->members->find(auth()->id());
+        if (!$actor->pivot->is_owner) {
+            abort(403);
+        }
+
         $groupInvitation->delete();
 
         return "ok";
